add missing TestLib package declaration
The TestLib module was missing a package declaration. Various test
scripts were loading the module and trying to import subs from it. Since
TestLib had no subs or any import method defined, those attempted
imports would be silently ignored. But because TestLib didn't declate a
package at all, its subs would end up in the current package of where it
was loaded, and they would be available anyway. This was not the
intended behavior, and future perl versions are intending to make it an
error to pass arguments to an undefined import method.